requirements.txt
transformers==4.46.2
torch==2.2.0
fastapi==0.105.0
uvicorn==0.30.0
gunicorn==21.2.0
python-dotenv==1.0.1